Obtain per-mesh settings from their settings classes in Mesh
We still have to do something about extruder index settings that are now obtained as size_t, while they might be -1. I'll look through the diff of this commit to get at those and obtain them via the ExtruderTrain& type instead, which does handle the -1 case.

Contributes to issue CURA-4410.

+++ src/FffPolygonGenerator.h
@@ -157,7 +157,7 @@ private:
      * \param mesh Input and Output parameter: fetches the outline information (see SliceLayerPart::outline) and generates the other reachable field of the \p storage
      * \param layer_nr The layer for which to generate the insets.
      */
-    void processInsets(const SliceDataStorage& storage, SliceMeshStorage& mesh, unsigned int layer_nr);
+    void processInsets(const SliceDataStorage& storage, SliceMeshStorage& mesh, size_t layer_nr);
 
     /*!
      * Generate the outline of the ooze shield.
